major changes to the architecture of the engine maps are split into 4 data files the .world format has changed to use the new maps the game singleton has built in editors new map loaders and saving code added etc

+++ 48HourFeb2009MiniRPGJamEntry/GameNameEntry.cpp
@@ -63,7 +63,7 @@ namespace GAME
 			const int numChar = std::strlen(keyMap_);
 			if(cursor_ == numChar) return 1;
 			if(cursor_ == numChar + 1) return -1;
-			if(name_.length() < maxChars_)
+			if(name_.length() < static_cast<unsigned int>(maxChars_))
 			{
 				name_ += keyMap_[cursor_];
 			}
